refactor(contracts): move governance error classes to governance/errors.ts
