+++ packages/contracts/src/governance/errors.ts
@@ -1,0 +1,48 @@
+/*
+ * This file is part of midnight-js.
+ * Copyright (C) 2025-2026 Midnight Foundation
+ * SPDX-License-Identifier: Apache-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { FinalizedTxData } from '@midnight-ntwrk/midnight-js-types';
+
+import { TxFailedError } from '../errors';
+
+/**
+ * An error indicating that a contract maintenance authority replacement transaction failed.
+ */
+export class ReplaceMaintenanceAuthorityTxFailedError extends TxFailedError {
+  constructor(finalizedTxData: FinalizedTxData) {
+    super(finalizedTxData);
+    this.name = 'ReplaceMaintenanceAuthorityTxFailedError';
+  }
+}
+
+/**
+ * An error indicating that a verifier key removal transaction failed.
+ */
+export class RemoveVerifierKeyTxFailedError extends TxFailedError {
+  constructor(finalizedTxData: FinalizedTxData) {
+    super(finalizedTxData);
+    this.name = 'RemoveVerifierKeyTxFailedError';
+  }
+}
+
+/**
+ * An error indicating that a verifier key insertion transaction failed.
+ */
+export class InsertVerifierKeyTxFailedError extends TxFailedError {
+  constructor(finalizedTxData: FinalizedTxData) {
+    super(finalizedTxData);
+    this.name = 'InsertVerifierKeyTxFailedError';
+  }
+}
